# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DEPOT API",
        "contact": {
            "name": "oussema khlifi",
            "email": "khoussemaa@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/depots/all": {
            "get": {
                "tags": [
                    "Depots"
                ],
                "summary": "Return the list of depots",
                "operationId": "6296d212bb45500fc483296f17b966c5",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/DepotSchema"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DepotSchema"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/depots/create": {
            "post": {
                "tags": [
                    "Depots"
                ],
                "summary": "Depot Create",
                "operationId": "cae79e0c87e7291556cc6306d6c2924a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "xxx"
                                ],
                                "properties": {
                                    "location": {
                                        "type": "string"
                                    },
                                    "size": {
                                        "type": "string"
                                    },
                                    "capacity": {
                                        "type": "integer"
                                    },
                                    "type": {
                                        "type": "string"
                                    },
                                    "isRented": {
                                        "type": "boolean"
                                    },
                                    "rent": {
                                        "type": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DepotSchema"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/depots/depot/{id}": {
            "get": {
                "tags": [
                    "Depots"
                ],
                "summary": "Depot Show",
                "operationId": "c2dc9afaff5f99bf2b18623d10be9789",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DepotSchema"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DepotSchema"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/depots/update/{id}": {
            "put": {
                "tags": [
                    "Depots"
                ],
                "summary": "Depot update",
                "operationId": "b2ebd4b33613925a47e7790245dd98b9",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "location": {
                                        "type": "string"
                                    },
                                    "size": {
                                        "type": "string"
                                    },
                                    "capacity": {
                                        "type": "integer"
                                    },
                                    "type": {
                                        "type": "string"
                                    },
                                    "isRented": {
                                        "type": "boolean"
                                    },
                                    "rent": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DepotSchema"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/depots/delete/{id}": {
            "delete": {
                "tags": [
                    "Depots"
                ],
                "summary": "Depot Delete",
                "operationId": "ba348ee72e03f27c06ca3d577fb812cf",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "DepotSchema": {
                "title": "Depot Model",
                "description": "Depot Model",
                "properties": {
                    "id": {
                        "description": "depot primary key",
                        "type": "uint64",
                        "example": "1"
                    },
                    "location": {
                        "description": "depot location (country, city, address)",
                        "type": "string",
                        "example": "tunis, sfax route tanyour km6"
                    },
                    "size": {
                        "description": "depot size",
                        "type": "string",
                        "example": "120x40m²"
                    },
                    "capacity": {
                        "description": "depot capacity of storing",
                        "type": "integer",
                        "example": "500"
                    },
                    "type": {
                        "description": "depot products stored type",
                        "type": "string",
                        "example": "food"
                    },
                    "isRented": {
                        "description": "depot rented verifier",
                        "type": "boolean",
                        "example": "1"
                    },
                    "rent": {
                        "description": "depot renting cost",
                        "type": "float",
                        "example": "1520.47"
                    }
                },
                "type": "object"
            },
...
        },
        "parameters": {
            "id": {
                "name": "id",
                "in": "query",
                "description": "depot id",
                "schema": {
                    "type": "number"
                }
            }
        }
    }
}